requirements.txt
Flask==2.3.2
google-generativeai==0.2.2